Split out items into a module
